+++ build.gradle
@@ -1,5 +1,6 @@
 
 dependencies {
+
 	implementation 'com.fasterxml.jackson.core:jackson-annotations:2.19.0'
 	implementation 'com.github.ulisesbocchio:jasypt-spring-boot-starter:3.0.5'
 }
