+++ src/main/java/minu/coffee/shop/ShopDto.java
@@ -1,0 +1,10 @@
+package minu.coffee.shop;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class ShopDto {
+
+
+
+}
